Utilize AOSP's new Volley http caching library
In order to maintain both the standalone app and the inline
app building I was forced to use another image caching library.
Since Volley is being announced at #IO13 this week it seems
the perfect fit. The library is slightly more complex and doesn't
provide some things like a default image while downloading but
it is very fast.

This commit just points to the new library allowing inline building.

Change-Id: Ie7882521f68f04a1b30f415a164a31178c2ea738